get_db_instrument and tests
